+++ src/lib_config.ml
@@ -4,4 +4,5 @@ type t =
   { has_native : bool
   ; ext_lib : string
   ; ext_obj : string
+  ; os_type : string
   }