+++ CHANGES.md
@@ -68,6 +68,9 @@ unreleased
 - Add the `%{os_type}` variable, which is a short-hand for
   `%{ocaml-config:os_type}` (#1764, @diml)
 
+- Allow `enabled_if` fields in `library` stanzas, restricted to the
+  `%{os_type}` variable (#1764, @diml)
+
 1.9.3 (06/05/2019)
 ------------------